pyproject.toml
[tool.poetry]
name = "Web-Project"
version = "0.1.0"
description = ""
authors = ["sazzeck <sazzeckprog@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.6"
python-dotenv = "^0.20.0"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
